Added average and p-p measurements

+++ scopehal/CaptureChannel.h
@@ -59,7 +59,7 @@ public:
 	/**
 		@brief Phase offset, in picoseconds, from the trigger to the sampling clock.
 	 */
-	int64_t m_triggerPhase;
+	double m_triggerPhase;
 
 	virtual size_t GetDepth() const =0;
